+++ src/main/java/frc/robot/RobotId.java
@@ -2,6 +2,7 @@ package frc.robot;
 
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.subsystems.drive.Drivetrain;
+import frc.robot.subsystems.drive.Module;
 
 import java.util.List;
 
@@ -13,7 +14,7 @@ import java.util.List;
  */
 public enum RobotId {
     Default,
-    SwerveCompetition(Drivetrain.class), SwerveTest(Drivetrain.class),
+    SwerveCompetition(Drivetrain.class, Module.class), SwerveTest(Drivetrain.class),
     ClassBot1, ClassBot2, ClassBot3, ClassBot4;
 
     /**
