+++ src/main/java/frc/robot/controls/GameControllerDriverConfig.java
@@ -7,14 +7,11 @@ import edu.wpi.first.wpilibj.shuffleboard.ShuffleboardTab;
 import edu.wpi.first.wpilibj2.command.InstantCommand;
 import frc.robot.commands.SetFormationX;
 import frc.robot.constants.miscConstants.OIConstants;
-import frc.robot.subsystems.DrivetrainImpl;
-import frc.robot.subsystems.Module;
+import frc.robot.subsystems.drive.DrivetrainImpl;
 import frc.robot.util.MathUtils;
 import lib.controllers.GameController;
 import lib.controllers.GameController.Axis;
 import lib.controllers.GameController.Button;
-
-import java.util.Arrays;
 
 /**
  * Driver controls for the generic game controller.
