+++ src/main/java/frc/robot/controls/PS5ControllerDriverConfig.java
@@ -8,14 +8,11 @@ import edu.wpi.first.wpilibj2.command.InstantCommand;
 import edu.wpi.first.wpilibj2.command.RepeatCommand;
 import frc.robot.commands.SetFormationX;
 import frc.robot.constants.miscConstants.OIConstants;
-import frc.robot.subsystems.DrivetrainImpl;
-import frc.robot.subsystems.Module;
+import frc.robot.subsystems.drive.DrivetrainImpl;
 import frc.robot.util.MathUtils;
 import lib.controllers.PS5Controller;
 import lib.controllers.PS5Controller.PS5Axis;
 import lib.controllers.PS5Controller.PS5Button;
-
-import java.util.Arrays;
 
 /**
  * Driver controls for the PS5 controller
