+++ src/main/java/frc/robot/subsystems/drive/Drivetrain.java
@@ -1,6 +1,8 @@
 package frc.robot.subsystems.drive;
 
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
+import frc.robot.constants.swerve.ModuleConstants;
+import frc.robot.subsystems.SubsystemFactory;
 import frc.robot.subsystems.SubsystemImpl;
 
 /**
@@ -10,6 +12,17 @@ import frc.robot.subsystems.SubsystemImpl;
  */
 @SubsystemImpl(DrivetrainImpl.class)
 public class Drivetrain extends SubsystemBase {
+
+    protected final Module[] modules;
+
+    public Drivetrain() {
+        modules = new Module[]{
+                SubsystemFactory.get(Module.class, ModuleConstants.FRONT_LEFT),
+                SubsystemFactory.get(Module.class, ModuleConstants.FRONT_RIGHT),
+                SubsystemFactory.get(Module.class, ModuleConstants.BACK_LEFT),
+                SubsystemFactory.get(Module.class, ModuleConstants.BACK_RIGHT)
+        };
+    }
 
     /**
      * Method to drive the robot using joystick info.
@@ -22,6 +35,10 @@ public class Drivetrain extends SubsystemBase {
      */
     public void drive(double xSpeed, double ySpeed, double rot, boolean fieldRelative, boolean isOpenLoop) {
         // TODO: Implement what this method should do as a placeholder
+    }
+
+    public Module[] getModules() {
+        return modules;
     }
 
 }
